MenuBar separated from Gui code
